course topic modeladmin changed to snippets
