+++ courses/admin.py
@@ -5,8 +5,6 @@ Admin site bindings for profiles
 from django.contrib import admin
 from django.db import models
 from django.forms import TextInput
-from wagtail.contrib.modeladmin.options import ModelAdmin, modeladmin_register
-from wagtail.contrib.modeladmin.views import DeleteView
 
 from mitxpro.admin import AuditableModelAdmin, TimestampedModelAdmin
 from mitxpro.utils import get_field_names
@@ -19,7 +17,6 @@ from .models import (
     CourseRunEnrollmentAudit,
     CourseRunGrade,
     CourseRunGradeAudit,
-    CourseTopic,
     Platform,
     Program,
     ProgramCertificate,
@@ -355,26 +352,6 @@ class ProgramCertificateAdmin(TimestampedModelAdmin):
         return self.model.all_objects.get_queryset().select_related("user", "program")
 
 
-class TopicsWagtailDeleteView(DeleteView):
-    """Custom view for Topics admin in Wagtail"""
-
-    def confirmation_message(self):
-        child_count = self.instance.subtopics.count()
-        if child_count > 0:
-            return (
-                f"This topic has {child_count} sub-topic(s) that will be deleted as well. Are you sure you want to "
-                f"delete? "
-            )
-        return "Are you sure you want to delete this topic?"
-
-
-class CourseTopicAdmin(ModelAdmin, admin.ModelAdmin):
-    """Admin for CourseTopic"""
-
-    model = CourseTopic
-    delete_view_class = TopicsWagtailDeleteView
-
-
 class PlatformAdmin(TimestampedModelAdmin):
     """Admin for Platform"""
 
@@ -383,7 +360,6 @@ class PlatformAdmin(TimestampedModelAdmin):
     search_fields = ["name"]
 
 
-modeladmin_register(CourseTopicAdmin)
 admin.site.register(Program, ProgramAdmin)
 admin.site.register(ProgramRun, ProgramRunAdmin)
 admin.site.register(Course, CourseAdmin)
