+++ pyproject.toml
@@ -36,7 +36,7 @@ celery-redbeat = "2.0.0"
 dj-database-url = "0.5.0"
 django = "3.2.23"
 django-anymail = { version = "8.4", extras = ["mailgun"] }
-django-filter = "2.4.0"
+django-filter = "^23.4"
 django-hijack = "2.1.10"
 django-hijack-admin = "2.1.10"
 django-ipware = "3.0.7"
@@ -77,7 +77,7 @@ ua-parser = "0.8.0"
 user-agents = "2.0"
 user-util = "0.1.5"
 uwsgi = "2.0.22"
-wagtail = "4.0.4"
+wagtail = "5.2"
 wagtail-metadata = "4.0.2"
 zeep = "3.4.0"
 flaky = "3.7.0"
